+++ python/mcap/mcap/data_stream.py
