+++ python/mcap/mcap/mcap0/exceptions.py
@@ -1,11 +1,1 @@
-class McapError(Exception):
-    pass
-
-
-class InvalidMagic(McapError):
-    def __init__(self):
-        super().__init__("not a valid MCAP file")
-
-
-class EndOfFile(McapError):
-    pass
+from ..exceptions import *
