+++ python/mcap/mcap/opcode.py
